OSIDB-4243: Add django-kaminarimon dependency and replace krb5_auth
This adds support for browser cookies for retrieving access tokens.

Closes OSIDB-4243
