+++ .secrets.baseline
@@ -470,10 +470,10 @@
         "filename": "tox.ini",
         "hashed_secret": "5a4fe08359c7f97380e408c717ef42c86939cd86",
         "is_verified": false,
-        "line_number": 54,
+        "line_number": 50,
         "is_secret": false
       }
     ]
   },
-  "generated_at": "2025-05-21T09:16:55Z"
+  "generated_at": "2025-06-02T15:10:23Z"
 }
