+++ config/urls.py
@@ -1,20 +1,20 @@
 """
 URL Configuration
 """
+
 from django.conf import settings
 from django.conf.urls.static import static
 from django.urls import include, path
-from rest_framework_simplejwt.views import TokenRefreshView, TokenVerifyView
+from kaminarimon.views import krb5_obtain_token_pair_view, refresh_token
+from rest_framework_simplejwt.views import TokenVerifyView
 
-from krb5_auth.views import OsidbTokenObtainPairView
+from osidb.api_views import OsidbTokenObtainPairView
 from osidb.views import index
 
 token_obtain_path = path(
     "auth/token", OsidbTokenObtainPairView.as_view(), name="token_obtain"
 )
-if "krb5_auth" in settings.INSTALLED_APPS:
-    from krb5_auth.views import krb5_obtain_token_pair_view
-
+if "kaminarimon" in settings.INSTALLED_APPS:
     token_obtain_path = path(
         "auth/token", krb5_obtain_token_pair_view, name="token_obtain"
     )
@@ -34,6 +34,6 @@ urlpatterns = [
     path("trackers/", include("apps.trackers.urls")),
     # auth
     token_obtain_path,
-    path("auth/token/refresh", TokenRefreshView.as_view(), name="token_refresh"),
+    path("auth/token/refresh", refresh_token, name="token_refresh"),
     path("auth/token/verify", TokenVerifyView.as_view(), name="token_verify"),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
