+++ tox.ini
@@ -37,10 +37,6 @@ commands =
 commands =
         pytest --record-mode=rewrite {posargs}
 
-[testenv:krb5-auth]
-commands =
-        pytest krb5_auth/
-
 [testenv:rls]
 deps = -rdevel-requirements.txt
        -rrequirements.txt
